Moved //hadoop/formats/cql one level up
