+++ hadoop/cql/CqlBinaryRecordReader.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.hadoop.formats.cql;
+package grakn.core.hadoop.cql;
 
 import com.datastax.oss.driver.api.core.cql.Row;
 import grakn.core.graph.diskstorage.Entry;
